! B fix parsing with key inside iterable
