+++ .dart_tool/package_config_subset
@@ -1,89 +1,101 @@
 async
 2.18
-file:///Users/art/.pub-cache/hosted/pub.dev/async-2.11.0/
-file:///Users/art/.pub-cache/hosted/pub.dev/async-2.11.0/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/async-2.11.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/async-2.11.0/lib/
 boolean_selector
 2.17
-file:///Users/art/.pub-cache/hosted/pub.dev/boolean_selector-2.1.1/
-file:///Users/art/.pub-cache/hosted/pub.dev/boolean_selector-2.1.1/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/boolean_selector-2.1.1/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/boolean_selector-2.1.1/lib/
 characters
 2.12
-file:///Users/art/.pub-cache/hosted/pub.dev/characters-1.3.0/
-file:///Users/art/.pub-cache/hosted/pub.dev/characters-1.3.0/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/characters-1.3.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/characters-1.3.0/lib/
 clock
 2.12
-file:///Users/art/.pub-cache/hosted/pub.dev/clock-1.1.1/
-file:///Users/art/.pub-cache/hosted/pub.dev/clock-1.1.1/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/clock-1.1.1/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/clock-1.1.1/lib/
 collection
-2.18
-file:///Users/art/.pub-cache/hosted/pub.dev/collection-1.18.0/
-file:///Users/art/.pub-cache/hosted/pub.dev/collection-1.18.0/lib/
+3.4
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/collection-1.19.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/collection-1.19.0/lib/
 fake_async
 2.12
-file:///Users/art/.pub-cache/hosted/pub.dev/fake_async-1.3.1/
-file:///Users/art/.pub-cache/hosted/pub.dev/fake_async-1.3.1/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/fake_async-1.3.1/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/fake_async-1.3.1/lib/
+leak_tracker
+3.2
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/leak_tracker-10.0.7/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/leak_tracker-10.0.7/lib/
+leak_tracker_flutter_testing
+3.2
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/leak_tracker_flutter_testing-3.0.8/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/leak_tracker_flutter_testing-3.0.8/lib/
+leak_tracker_testing
+3.2
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/leak_tracker_testing-3.0.1/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/leak_tracker_testing-3.0.1/lib/
 matcher
-2.18
-file:///Users/art/.pub-cache/hosted/pub.dev/matcher-0.12.16/
-file:///Users/art/.pub-cache/hosted/pub.dev/matcher-0.12.16/lib/
+3.0
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/matcher-0.12.16+1/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/matcher-0.12.16+1/lib/
 material_color_utilities
 2.17
-file:///Users/art/.pub-cache/hosted/pub.dev/material_color_utilities-0.5.0/
-file:///Users/art/.pub-cache/hosted/pub.dev/material_color_utilities-0.5.0/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/material_color_utilities-0.11.1/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/material_color_utilities-0.11.1/lib/
 meta
 2.12
-file:///Users/art/.pub-cache/hosted/pub.dev/meta-1.10.0/
-file:///Users/art/.pub-cache/hosted/pub.dev/meta-1.10.0/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/meta-1.15.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/meta-1.15.0/lib/
 path
-2.12
-file:///Users/art/.pub-cache/hosted/pub.dev/path-1.8.3/
-file:///Users/art/.pub-cache/hosted/pub.dev/path-1.8.3/lib/
+3.0
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/path-1.9.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/path-1.9.0/lib/
 source_span
 2.18
-file:///Users/art/.pub-cache/hosted/pub.dev/source_span-1.10.0/
-file:///Users/art/.pub-cache/hosted/pub.dev/source_span-1.10.0/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/source_span-1.10.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/source_span-1.10.0/lib/
 stack_trace
-2.18
-file:///Users/art/.pub-cache/hosted/pub.dev/stack_trace-1.11.1/
-file:///Users/art/.pub-cache/hosted/pub.dev/stack_trace-1.11.1/lib/
+3.4
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/stack_trace-1.12.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/stack_trace-1.12.0/lib/
 stream_channel
 2.19
-file:///Users/art/.pub-cache/hosted/pub.dev/stream_channel-2.1.2/
-file:///Users/art/.pub-cache/hosted/pub.dev/stream_channel-2.1.2/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/stream_channel-2.1.2/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/stream_channel-2.1.2/lib/
 string_scanner
-2.18
-file:///Users/art/.pub-cache/hosted/pub.dev/string_scanner-1.2.0/
-file:///Users/art/.pub-cache/hosted/pub.dev/string_scanner-1.2.0/lib/
+3.1
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/string_scanner-1.3.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/string_scanner-1.3.0/lib/
 term_glyph
 2.12
-file:///Users/art/.pub-cache/hosted/pub.dev/term_glyph-1.2.1/
-file:///Users/art/.pub-cache/hosted/pub.dev/term_glyph-1.2.1/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/term_glyph-1.2.1/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/term_glyph-1.2.1/lib/
 test_api
-3.0
-file:///Users/art/.pub-cache/hosted/pub.dev/test_api-0.6.1/
-file:///Users/art/.pub-cache/hosted/pub.dev/test_api-0.6.1/lib/
+3.4
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/test_api-0.7.3/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/test_api-0.7.3/lib/
 vector_math
 2.14
-file:///Users/art/.pub-cache/hosted/pub.dev/vector_math-2.1.4/
-file:///Users/art/.pub-cache/hosted/pub.dev/vector_math-2.1.4/lib/
-web
-3.2
-file:///Users/art/.pub-cache/hosted/pub.dev/web-0.3.0/
-file:///Users/art/.pub-cache/hosted/pub.dev/web-0.3.0/lib/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/vector_math-2.1.4/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/vector_math-2.1.4/lib/
+vm_service
+3.3
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/vm_service-14.3.0/
+file:///Users/yelamanyelmuratov/.pub-cache/hosted/pub.dev/vm_service-14.3.0/lib/
 dart_safe_data_class_test
 3.2
-file:///Users/art/Documents/GitHub/dart-data-class-tools/
-file:///Users/art/Documents/GitHub/dart-data-class-tools/lib/
+file:///Users/yelamanyelmuratov/Development/extensions/dart-data-class-tools/
+file:///Users/yelamanyelmuratov/Development/extensions/dart-data-class-tools/lib/
 sky_engine
 3.2
-file:///Users/art/flutter/bin/cache/pkg/sky_engine/
-file:///Users/art/flutter/bin/cache/pkg/sky_engine/lib/
+file:///opt/homebrew/Caskroom/flutter/3.27.2/flutter/bin/cache/pkg/sky_engine/
+file:///opt/homebrew/Caskroom/flutter/3.27.2/flutter/bin/cache/pkg/sky_engine/lib/
 flutter
-3.2
-file:///Users/art/flutter/packages/flutter/
-file:///Users/art/flutter/packages/flutter/lib/
+3.3
+file:///opt/homebrew/Caskroom/flutter/3.27.2/flutter/packages/flutter/
+file:///opt/homebrew/Caskroom/flutter/3.27.2/flutter/packages/flutter/lib/
 flutter_test
-3.2
-file:///Users/art/flutter/packages/flutter_test/
-file:///Users/art/flutter/packages/flutter_test/lib/
+3.3
+file:///opt/homebrew/Caskroom/flutter/3.27.2/flutter/packages/flutter_test/
+file:///opt/homebrew/Caskroom/flutter/3.27.2/flutter/packages/flutter_test/lib/
 2
