+++ example/README.md
@@ -1,0 +1,3 @@
+# example
+
+A new Flutter project.
