--- pubspec.yaml
@@ -1,17 +1,0 @@
-name: dart_safe_data_class_test
-description: test
-publish_to: 'none'
-version: 0.1.0
-
-environment:
-  sdk: '>=3.2.0 <4.0.0'
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-
-flutter:
